+++ webview/src/pages/WebviewPage.tsx
@@ -15,8 +15,8 @@ import { parseOpenAPIV2 } from '@/utils/parseSwaggerDocs';
 import { ApiGroupByTag, ApiPathType } from '@/utils/types';
 import { DownOutlined, FolderAddOutlined, LinkOutlined, SettingOutlined, UploadOutlined } from '@ant-design/icons';
 import { usePromisifyModal } from '@orca-fe/hooks';
-import { useBoolean, useDebounceEffect, useMap, useMemoizedFn, useMount, useToggle } from 'ahooks';
-import { Affix, Button, Collapse, Empty, FloatButton, Form, Layout, Modal, Space, Spin, Tabs, Tooltip, Typography, Upload, theme } from 'antd';
+import { useBoolean, useDebounceEffect, useDebounceFn, useMap, useMemoizedFn, useMount, useToggle } from 'ahooks';
+import { Affix, Button, Checkbox, Collapse, Empty, FloatButton, Form, Layout, Modal, Space, Spin, Tabs, Tooltip, Typography, Upload, theme } from 'antd';
 import { OpenAPIV2 } from 'openapi-types';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import styles from './WebviewPage.less';
@@ -31,6 +31,8 @@ import { WebviewPageContext } from './context';
 import SettingModal from '@/components/SettingModal';
 import { pick } from 'lodash-es';
 import SearchSuite, { SearchValue } from '@/components/SearchSuite';
+import { ADVANCED_SEARCH_OPTIONS, PARSE_METHOD_DOCS, PARSE_METHOD_LOCAL, SEARCH_FILTER } from './constants';
+import { CheckboxValueType } from 'antd/es/checkbox/Group';
 
 const { Header, Content } = Layout;
 const { useForm, useWatch } = Form;
@@ -40,9 +42,6 @@ const formItemLayout = {
   labelCol: { span: 24 },
   wrapperCol: { span: 18 },
 };
-
-const PARSE_METHOD_DOCS = 'swagger docs';
-const PARSE_METHOD_LOCAL = 'local file';
 
 export interface WebviewPageProps extends React.HTMLAttributes<HTMLDivElement> {}
 
@@ -54,6 +53,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
   const [currentApiGroup, setCurrentApiGroup] = useState<ApiGroupByTag[]>([]);
   const [swaggerDocs, setSwaggerDocs] = useState<OpenAPIV2.Document>();
   const [searchPanelKey, setSearchPanelKey] = useState<string>(PARSE_METHOD_DOCS);
+  const [filters, setFilters] = useState<CheckboxValueType[]>([SEARCH_FILTER.HIDE_EMPTY_GROUP]);
   const [openApiPanelKeys, setOpenApiPanelKeys] = useState<string[]>([]);
   const _this = useRef<{ V2Document?: OpenAPIV2.Document }>({}).current;
 
@@ -117,7 +117,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
     resetPageWhenChange();
     const swaggerDocs = resp.data;
     if (!resp.success || !resp.data) {
-      notification.error(resp.errMsg ?? 'Swagger 文档解析失败, 请稍后再试');
+      notification.error(resp.errMsg ?? 'Swagger 文档解析失败，请稍后再试');
       return;
     }
     if (!('swagger' in swaggerDocs)) {
@@ -125,7 +125,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
       return;
     }
     const currentApiName = options.find((option) => option.value === currentSwaggerUrl)?.label;
-    notification.success(`【${currentApiName}】 Swagger 文档解析成功`);
+    notification.success(`【${currentApiName}】Swagger 文档解析成功`);
     handleV2DocumentData(swaggerDocs as OpenAPIV2.Document);
   });
 
@@ -305,7 +305,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
                           name="swaggerUrl"
                           required
                           rules={[{ required: true, message: '请选择一个 Swagger 地址' }]}
-                          style={{ marginTop: 4 }}
+                          style={{ marginTop: 4, marginBottom: 0 }}
                           label={
                             <Space>
                               <span>文档接口地址：</span>
@@ -350,6 +350,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
                     </Tooltip>
                   }
                 />
+                  <Checkbox.Group value={filters} onChange={setFilters} options={ADVANCED_SEARCH_OPTIONS} style={{ margin: '8px 0 12px 0'}}></Checkbox.Group>
                 <Form.Item
                   name="searchParams"
                   label={
@@ -390,7 +391,7 @@ const WebviewPage: React.FC<WebviewPageProps> = (props) => {
             </Content>
           </Affix>
           <Spin spinning={parseLoading}>
-            <WebviewPageContext.Provider value={{ refreshDocFlag }}>
+            <WebviewPageContext.Provider value={{ refreshDocFlag, filters }}>
               {hasSwaggerDocs && <SwaggerInfo className={styles.swaggerInfo} v2Doc={swaggerDocs} />}
               {!!currentApiGroup.length ? (
                 <Collapse