+++ webview/src/pages/context.ts
@@ -1,3 +1,4 @@
+import { CheckboxValueType } from 'antd/es/checkbox/Group';
 import React from 'react';
 
-export const WebviewPageContext = React.createContext<{ refreshDocFlag?: boolean }>({});
+export const WebviewPageContext = React.createContext<{ refreshDocFlag?: boolean, filters?: CheckboxValueType[] }>({});